pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "campaign-manager"
version = "0.1.0"
description = "CLI tool for managing marketing campaigns across Reddit, Dev.to, and more"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "praw>=7.7",
    "requests>=2.31",
    "python-dotenv>=1.0",
    "rich>=13.0",
]

[project.scripts]
marketing = "marketing.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]
